// src/enums/BookingEnum.ts
export enum BookingStatusEnum {
    FAILED = "FAILED",
    IN_PROGRESS = "IN_PROGRESS",
    RETRY_COMPLETED = "RETRY_COMPLETED",
    RETRY_FAILED = "RETRY_FAILED",
    STARTED = "STARTED",
    BACKLOG = "BACKLOG",
}

export enum BookingServiceEnum {
    BOOKING_CREATED = "BOOKING_CREATED",
    BOOKING_EMAIL = "BOOKING_EMAIL",
    BOOKING_TASK = "BOOKING_TASK",
    BOOKING_NOTIFICATION = "BOOKING_NOTIFICATION",
    BOOKING_LOGS = "BOOKING_LOGS",
    BOOKING_ADMIN_CUSTOMER = "BOOKING_ADMIN_CUSTOMER",
}